JavaScript: The Bad Parts
Previously we were using JSLint as a parser, which only supports a very
strict subset of JavaScript. However, there are times when you want to
quickly evaluate a very sloppy ad hoc expression, or a 3rd party module
which uses bad parts of the language.

This commit adds support for the subset of JavaScript which Acorn
supports, which is considerably broader. I am hoping that this change
will make Replete more accessible.