Deno
This commit adds support for Deno, a less crufty alternative to Node.js.

This required a radical restructuring of the Node.js REPL: previously,
source was evaluated in the same process as the REPL.  Now, source is
evaluated in a seperate process. This is good because the evaluated
source can now do things like:

- manipulate the global object
- kill the process
- write to STDOUT/STDERR

without affecting the REPL's process. We had to use the experimental
"module loader" feature, so support for Node.js versions below 17 have
been dropped.