Always array
Previously, the browser REPL produced an array of evaluations, whereas
the other REPLs produced a single evaluation. This introduced an
ambiguity when wun of the other REPLs produced an array: was the array
the evaluation, or does it contain the evaluations?

This commit resolves the issue by always producing an array of
evaluations, even if there will only ever be wun.

This commit also makes a change to the WEBL client. Upon restarting
Replete (and hence the WEBL server), the WEBL client will reload the
page, thereby clearing the module cache. I have found this to be a much
better experience when working on multiple modules.