Bidirectional messages
Previously, replete.js emitted an unstructured text stream to STDOUT and
STDERR. I chose this behaviour because it made it easy for the plugin to
dump that text into a view.

However, it seems likely that plugin makers will want to distinguish
between the different types of output so that they can be displayed
differently.  This commit changes replete.js so that it emits JSON
messages instead of unstructured text.