Result messages
Previously, replete.js sent back messages of the form {type, string}.
It now sends back result messages with a single property containing the
string. The property name is wun of

    evaluation
    exception
    out
    err

which I think is neater.

I also learned that passing a second function to 'then' is not the same
as passing a function to 'catch'. The 'catch' catches rejections from
the entire Promise chain, which is what I meant to do. The process no
longer crashes when the 'read' capability fails.