URL locators
Previously, some locators began with a "/" to indicate that they
referred to a file on disk. This was a mistake, because paths on Windows
do not start with a forward slash. A whole demographic of programmers
was excluded from enjoying the goodness of Replete.

This commit changes the definition of a locator such that all locators
are now URLs. For example, the path "C:\path\to\file.js" can now be
represented as a locator like "file:///C:/path/to/file.js". This is a
standard transformation, and Node.js can read such a URL directly on
Windows systems.

A benefit of this change is that all locators are now URLs, which is
simpler than the old path/URL duality.

This change will lead to breakage in custom capability functions and
text editor plugins.