$
This commit changes how function declarations are handled by the REPL.
Previously, we prefixed function names with "$" to avoid shadowing their
outer variable. As well as being unsightly and somewhat confusing, this
was also slightly dangerous for function names already beginning with $.

We do away with the $, but in doing so we introduce a new caveat, which
is probably more dangerous.

Hopefully nobody ever notices it.