Sleep tolerance
Previously, when source was evaluated in multiple WEBL clients
concurrently, the results were reported all together. This is less than
ideal when wun of the WEBL browser tabs had gone to sleep or not cleanly
closed its WebSocket connection. Because it would never respond, the
results of the other WEBL clients would never be reported back either.

This commit changes the behaviour of the browser REPL. It now reports
back the result from each WEBL client as it arrives. Any malfunctioning
WEBL clients can be ignored.