import.meta
This commit adds support for the dynamic import function and the
import.meta object.

The import.meta object is used to resolve the address of files relative
to the current module. This makes it much easier to use Workers, which
are resolved relative to the page

    new Worker("/absolute/path/to/my_worker.js");

rather than the module

    new Worker(import.meta.resolve("./my_worker.js"));

Few runtimes currently support import.meta.resolve, so an approach using
import.meta.url is supported too:

    new Worker(new URL("./my_worker.js", import.meta.url));

Workers are available in the browser and Deno. They allow JavaScript
programs to parallelise their work over multiple CPU cores.