top
Previously, the browser REPL evaluated code within an iframe. This was because
the WEBL was designed to sandbox evaluation, but this is not a requirement for
REPLing.  Now, code is evaluated directly in the WEBL client, making it easier
for the programmer to locate elements within the browser's element inspector
panel.